Ridge object properly defined ! Cells are linked together
Correctly defined the Ridge object with getter and setters for
the up,down,left,right properties. These properties are linked to the
global common ridgeh/v arrays, making possible to have 2 cells pointing to the
same ridge. And so modifying a ridge of one cell automatically modify the ridge
of the adjacent cell.